+++ docs/thorest/thor/model/model.mmd
@@ -55,3 +55,8 @@ classDiagram
         constructor(TransferJSON json)
         TransferJSON toJSON()
     }
+    class TXID {
+        TxId id
+        constructor(TXIDJSON json)
+        TXIDJSON toJSON()
+    }
