+++ packages/thorest/src/thor/model/Output.ts
@@ -33,7 +33,7 @@ class Output {
      * Constructs an instance of the class using the provided OutputJSON object.
      *
      * @param {OutputJSON} json - The JSON object containing the required fields to initialize the instance.
-     * @throws {IllegalArgumentError} Throws an error if the JSON object cannot be parsed or contains invalid values.
+     * @throws {IllegalArgumentError} If the JSON object cannot be parsed or contains invalid values.
      */
     constructor(json: OutputJSON) {
         try {
@@ -58,7 +58,7 @@ class Output {
     }
 
     /**
-     * Converts the current instance of the class into a OutputJSON representation.
+     * Converts the current instance of the class into an OutputJSON representation.
      *
      * @return {OutputJSON} The JSON object representing the current instance.
      */
