+++ packages/thorest/src/thor/model/TXJSON.ts
@@ -1,0 +1,8 @@
+/**
+ * [TXID](http://localhost:8669/doc/stoplight-ui/#/schemas/TXID)
+ */
+interface TXIDJSON {
+    id: string;
+}
+
+export { type TXIDJSON };
