+++ packages/thorest/src/thor/transactions/index.ts
@@ -10,6 +10,5 @@ export * from './RetrieveRawTransactionByID';
 export * from './RetrieveTransactionByID';
 export * from './RetrieveTransactionReceipt';
 export * from './SendTransaction';
-export * from './TXID';
 export * from './TxMeta';
 export * from './TxMetaJSON';
